pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shieldpilot"
version = "0.2.0"
description = "Security platform protecting autonomous AI coding agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "ShieldPilot Team"}
]
keywords = ["security", "ai", "claude", "agent", "llm", "safety"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
]
dependencies = [
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "typer[all]>=0.9.0,<1.0.0",
    "rich>=13.0,<14.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "pydantic[email]>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "pyyaml>=6.0,<7.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "python-multipart>=0.0.6,<1.0.0",
    "aiofiles>=23.0,<24.0.0",
    "httpx>=0.25.0,<1.0.0",
    "stripe>=7.0.0,<11.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "alembic>=1.12.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "joblib>=1.3.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/sentinelai/sentinelai"
Documentation = "https://github.com/sentinelai/sentinelai#readme"
Repository = "https://github.com/sentinelai/sentinelai"
Issues = "https://github.com/sentinelai/sentinelai/issues"

[project.optional-dependencies]
llm = [
    "anthropic>=0.7.0",
]
ml = [
    "transformers>=4.30.0",
    "onnxruntime>=1.15.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
all = [
    "sentinelai[llm,ml,dev]",
]

[project.scripts]
sentinel = "sentinelai.cli.main:app"
sentinel-hook = "sentinelai.hooks.sentinel_hook:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["sentinelai*"]

[tool.setuptools.package-data]
sentinelai = [
    "web/templates/*.html",
    "web/static/css/*.css",
    "web/static/js/*.js",
    "migrations/*.py",
    "migrations/*.mako",
    "migrations/versions/*.py",
    "migrations/README",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unix_only: test requires Unix (Linux/macOS) shell",
    "linux_only: test requires Linux specifically",
]

[tool.ruff]
target-version = "py39"
line-length = 100
